requirements.txt
bcrypt==3.1.3
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
coverage==4.4.1
DateTime==4.2
Django==1.11.3
djangorestframework==3.6.3
gunicorn==19.7.1
idna==2.5
mysqlclient==1.3.10
pycparser==2.18
pytz==2017.2
requests==2.18.1
six==1.10.0
urllib3==1.21.1
zope.interface==4.4.2
